+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -319,7 +319,11 @@ class ExpoCameraView(
       .build()
 
     recorder?.let {
-      if (!mute && ActivityCompat.checkSelfPermission(context, Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
+      if (!mute && ActivityCompat.checkSelfPermission(
+          context,
+          Manifest.permission.RECORD_AUDIO
+        ) != PackageManager.PERMISSION_GRANTED
+      ) {
         promise.reject(Exceptions.MissingPermissions(Manifest.permission.RECORD_AUDIO))
         return
       }
@@ -354,7 +358,11 @@ class ExpoCameraView(
           }
         }
     }
-      ?: promise.reject("E_RECORDING_FAILED", "Starting video recording failed - could not create video file.", null)
+      ?: promise.reject(
+        "E_RECORDING_FAILED",
+        "Starting video recording failed - could not create video file.",
+        null
+      )
   }
 
   @SuppressLint("UnsafeOptInUsageError")
@@ -367,11 +375,12 @@ class ExpoCameraView(
       {
         val cameraProvider: ProcessCameraProvider = providerFuture.get()
 
-        previewView.scaleType = if (ratio == CameraRatio.FOUR_THREE || ratio == CameraRatio.SIXTEEN_NINE) {
-          PreviewView.ScaleType.FIT_CENTER
-        } else {
-          PreviewView.ScaleType.FILL_CENTER
-        }
+        previewView.scaleType =
+          if (ratio == CameraRatio.FOUR_THREE || ratio == CameraRatio.SIXTEEN_NINE) {
+            PreviewView.ScaleType.FIT_CENTER
+          } else {
+            PreviewView.ScaleType.FILL_CENTER
+          }
 
         val resolutionSelector = buildResolutionSelector()
         val preview = Preview.Builder()
@@ -456,8 +465,10 @@ class ExpoCameraView(
 
   private fun buildResolutionSelector(): ResolutionSelector {
     val strategy = if (pictureSize.isNotEmpty()) {
-      val size = Size.parseSize(pictureSize)
-      ResolutionStrategy(size, ResolutionStrategy.FALLBACK_RULE_CLOSEST_LOWER_THEN_HIGHER)
+      val size = parseSizeSafely(pictureSize)
+      size?.let {
+        ResolutionStrategy(size, ResolutionStrategy.FALLBACK_RULE_CLOSEST_LOWER_THEN_HIGHER)
+      } ?: ResolutionStrategy.HIGHEST_AVAILABLE_STRATEGY
     } else {
       ResolutionStrategy.HIGHEST_AVAILABLE_STRATEGY
     }
@@ -475,6 +486,19 @@ class ExpoCameraView(
         }
         setResolutionStrategy(strategy)
       }.build()
+    }
+  }
+
+  private fun parseSizeSafely(size: String): Size? {
+    val pattern = Regex("\\d+x\\d+")
+    if (!pattern.matches(size)) {
+      return null
+    }
+
+    return try {
+      Size.parseSize(size)
+    } catch (e: Throwable) {
+      null
     }
   }
 
@@ -511,7 +535,10 @@ class ExpoCameraView(
         previewView.width.toFloat(),
         previewView.height.toFloat()
       )
-      val action = FocusMeteringAction.Builder(meteringPointFactory.createPoint(1f, 1f), FocusMeteringAction.FLAG_AF)
+      val action = FocusMeteringAction.Builder(
+        meteringPointFactory.createPoint(1f, 1f),
+        FocusMeteringAction.FLAG_AF
+      )
         .build()
       it.cameraControl.startFocusAndMetering(action)
     }
@@ -533,7 +560,8 @@ class ExpoCameraView(
   @OptIn(ExperimentalCamera2Interop::class)
   fun getAvailablePictureSizes(): List<String> {
     return camera?.cameraInfo?.let { cameraInfo ->
-      val info = Camera2CameraInfo.from(cameraInfo).getCameraCharacteristic(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP)
+      val info = Camera2CameraInfo.from(cameraInfo)
+        .getCameraCharacteristic(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP)
       info?.getOutputSizes(ImageFormat.JPEG)?.map { it.toString() }
     } ?: emptyList()
   }
@@ -599,7 +627,10 @@ class ExpoCameraView(
     barcode.referenceImageWidth = width
   }
 
-  private fun getCornerPointsAndBoundingBox(cornerPoints: List<Int>, boundingBox: BoundingBox): Pair<ArrayList<Bundle>, Bundle> {
+  private fun getCornerPointsAndBoundingBox(
+    cornerPoints: List<Int>,
+    boundingBox: BoundingBox
+  ): Pair<ArrayList<Bundle>, Bundle> {
     val density = previewView.resources.displayMetrics.density
     val convertedCornerPoints = ArrayList<Bundle>()
     for (i in cornerPoints.indices step 2) {
@@ -634,7 +665,10 @@ class ExpoCameraView(
   private fun onBarcodeScanned(barcode: BarCodeScannerResult) {
     if (shouldScanBarcodes) {
       transformBarcodeScannerResultToViewCoordinates(barcode)
-      val (cornerPoints, boundingBox) = getCornerPointsAndBoundingBox(barcode.cornerPoints, barcode.boundingBox)
+      val (cornerPoints, boundingBox) = getCornerPointsAndBoundingBox(
+        barcode.cornerPoints,
+        barcode.boundingBox
+      )
       onBarcodeScanned(
         BarcodeScannedEvent(
           target = id,